[refactor] Migrate buttons to core module.
ボタンをcoreモジュールに移行
