+++ androidApp/core/build.gradle.kts
@@ -44,5 +44,4 @@ dependencies {
     implementation(libs.bundles.compose)
     implementation(libs.bundles.accompanist)
     implementation(libs.coil.compose)
-    implementation(projects.androidApp.core)
 }
