+++ androidApp/home/src/main/java/tech/takahana/iconwallpaper/android/home/ui/screen/SelectStuffScreen.kt
@@ -11,10 +11,10 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import tech.takahana.iconwallpaper.android.core.ui.components.PrimaryColorButton
 import tech.takahana.iconwallpaper.android.home.R
 import tech.takahana.iconwallpaper.android.home.ui.components.Announcements
 import tech.takahana.iconwallpaper.android.home.ui.components.ItemGrid
-import tech.takahana.iconwallpaper.android.home.ui.components.PrimaryColorButton
 
 @Composable
 fun SelectStuffScreen() {
