+++ androidApp/onbording/build.gradle.kts
@@ -44,4 +44,5 @@ dependencies {
     implementation(libs.bundles.compose)
     implementation(libs.bundles.accompanist)
     implementation(libs.coil.compose)
+    implementation(projects.androidApp.core)
 }
