+++ settings.gradle.kts
@@ -16,6 +16,7 @@ include(":uilogic:welcome")
 include(":usecase:onboarding")
 include(":uilogicImpl:welcome")
 include(":usecaseImpl:onboarding")
+include(":androidApp:core")
 include(":androidApp:onbording")
 include(":androidApp:home")
 include(":usecase:home")
